Fixed clone progress display.
The type of block_total/block_done was changed to int64 in
https://github.com/haiwen/seafile/commit/656666a#diff-2d685db41f30dc382661f0dfcf5804a9R27

+++ src/rpc/clone-task.h
@@ -19,8 +19,8 @@ public:
     QString tx_id;
 
     QString state_str;
-    int block_done;
-    int block_total;
+    qint64 block_done;
+    qint64 block_total;
 
     int fs_objects_done;
     int fs_objects_total;
